[be/fix] Long 으로 변환. 이에 따른 로직 에러 수정

+++ backend/notice/src/main/java/com/dokidoki/notice/kafka/dto/KafkaAuctionRegisterDTO.java
@@ -8,17 +8,17 @@ import lombok.*;
 @Builder
 public class KafkaAuctionRegisterDTO {
 
-    private long auctionId;
-    private int offerPrice;             // 시작 가격(호가)
-    private int priceSize;              // 경매 단위
-    private long ttl;
-    private long sellerId;
-    private long productId;
+    private Long auctionId;
+    private Long offerPrice;             // 시작 가격(호가)
+    private Long priceSize;              // 경매 단위
+    private Long ttl;
+    private Long sellerId;
+    private Long productId;
     private String productName;
 
     public KafkaAuctionRegisterDTO() {}
 
-    public KafkaAuctionRegisterDTO(long auctionId, int offerPrice, int priceSize, long ttl, long sellerId, long productId, String productName) {
+    public KafkaAuctionRegisterDTO(Long auctionId, Long offerPrice, Long priceSize, Long ttl, Long sellerId, Long productId, String productName) {
         this.auctionId = auctionId;
         this.offerPrice = offerPrice;
         this.priceSize = priceSize;
